rag/requirements.txt
# PDF 处理
pymupdf4llm==1.27.2.2
PyMuPDF==1.27.2.2

# 文本切分
langchain==1.2.0
langchain-text-splitters==1.0.0
langchain-core==1.1.3

# 向量数据库
qdrant-client==1.17.1

# 嵌入模型
dashscope==1.25.3

# 环境管理
python-dotenv==1.2.1

# 工具库
numpy==2.3.5
requests==2.32.5
tqdm==4.67.1
rich==14.3.3
